requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
torchvision
pillow
joblib
numpy
pynvml
